server/requirements.txt
Django>=5
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg[binary]
python-dotenv
pytest>=7.4
pytest-django>=4.8
